+++ crates/lsp/src/handlers_ext.rs
